moved away from NPolyglot
