+++ Release-Application.ps1
@@ -1,0 +1,2 @@
+dotnet publish src/RecordsLang -o ../../release -c Release
+nuget pack build/RecordsLang.nuspec -OutputDirectory build
