--- src/RecordsLang.Runner/RecordsLang.Runner.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="../RecordsLang/RecordsLang.csproj" />
-  </ItemGroup>
-
-</Project>
