--- test/RecordsLang.Test/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NPolyglot.Core" version="0.2.0-alpha" targetFramework="net452" />
-</packages>
